requirements.txt
flask==2.3.3
flask-cors==4.0.0
pymongo==4.5.0
openai==0.28.1
requests==2.31.0
reportlab==4.0.4
pdfkit==1.0.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
textstat==0.7.3
nltk==3.8.1
google-generativeai==0.3.2
PyPDF2==3.0.1
matplotlib==3.7.2
seaborn==0.12.2
numpy==1.24.3